Fix sorting broken by a limit on search for valid locations
In this context, "self" are sorted locations. The search returns valid
locations. Later, we filter the locations in "self" by keeping only the
valid locations, but keeping the order of "self.ids".  If we apply a
limit on the intermediate search for valid locations, the order of
"self.ids" may be wrong as the search will return only the first valid
id(s) according to its own order which may be different.